requirements.txt
azure-common==1.1.28
azure-core==1.29.5
azure-functions==1.17.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.18.3
azure-storage-common==2.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
idna==3.4
isodate==0.6.1
msal==1.25.0
msal-extensions==1.0.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
pytest==7.4.3